Redesigned making each project to inherit directly from Spring to enable projects to share functionality
