+++ .gitignore
@@ -19,6 +19,7 @@ repos
 .project
 .settings/
 .gradle
+.idea/
 
 # Mobile Tools for Java (J2ME)
 .mtj.tmp/
