+++ commons/src/main/java/org/springboardfoundation/common/entity/Gender.java
@@ -1,0 +1,5 @@
+package org.springboardfoundation.common.entity;
+
+public enum Gender {
+    male,female,others
+}
