--- commons/src/main/java/org/springboardfoundation/usersservice/entity/Gender.java
@@ -1,5 +1,0 @@
-package org.springboardfoundation.usersservice.entity;
-
-public enum Gender {
-    male,female,others
-}
