+++ db-lib/src/main/java/org/springbordfoundation/db/entity/UserTrigger.java
@@ -1,0 +1,24 @@
+package org.springbordfoundation.db.entity;
+
+import jakarta.persistence.*;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@Entity
+@Table(name = "user_trigger")
+public class UserTrigger {
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "id")
+    private Long id;
+
+    @OneToOne(cascade = CascadeType.ALL)
+    @JoinColumn(name = "user_identifier", referencedColumnName = "id")
+    private User user;
+
+    private String responseId;
+
+    private String responseCode;
+}
