+++ db-lib/src/main/java/org/springbordfoundation/db/respositories/UserRepositoy.java
@@ -1,9 +1,11 @@
 package org.springbordfoundation.db.respositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springboardfoundation.usersservice.entity.User;
+import org.springbordfoundation.db.entity.User;
+import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
 
+@Component
 @Repository
 public interface UserRepositoy extends JpaRepository<User,Long> {
