+++ db-lib/src/main/java/org/springbordfoundation/db/respositories/UserTriggerRepositoy.java
@@ -1,6 +1,6 @@
-package org.springboardfoundation.usersservice.repositories;
+package org.springbordfoundation.db.respositories;
 
-import org.springboardfoundation.usersservice.entity.UserTrigger;
+import org.springbordfoundation.db.entity.UserTrigger;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
