+++ db-lib/src/main/java/org/springbordfoundation/db/service/UserDbService.java
@@ -1,10 +1,12 @@
 package org.springbordfoundation.db.service;
 
-import org.springboardfoundation.usersservice.entity.User;
+import org.springbordfoundation.db.entity.User;
 
 public interface UserDbService {
 
     void save(User user);
+
+    void update(User user);
 
     User findByUserIdentifier(String userIdentifier);
