--- users-service/src/main/java/org/springboardfoundation/usersservice/repositories/UserRepositoy.java
@@ -1,11 +1,0 @@
-package org.springboardfoundation.usersservice.repositories;
-
-import org.springboardfoundation.usersservice.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepositoy extends JpaRepository<User,Long> {
-
-    User findByUserIdentifier(String userIdentifier);
-}
